MARS-1010 Create new `Data` routes
